+++ data_pipeline/s3_csv_data/cli.py
@@ -1,23 +1,13 @@
 import argparse
 import logging
-import os
 from typing import Optional, Sequence
 
 from data_pipeline.s3_csv_data.s3_csv_config import (
-    DEFAULT_INITIAL_S3_FILE_LAST_MODIFIED_DATE,
     MultiS3CsvConfig,
     S3BaseCsvConfig
 )
 from data_pipeline.s3_csv_data.s3_csv_etl import (
-    get_stored_state,
-    transform_load_data,
-    update_object_latest_dates,
-    upload_s3_object_json
-)
-from data_pipeline.utils.data_pipeline_timestamp import get_current_timestamp_as_string
-from data_pipeline.utils.data_store.s3_data_service import (
-    get_s3_object_etag,
-    iter_sorted_new_s3_files_to_process
+    etl_new_csv_files
 )
 from data_pipeline.utils.pipeline_config import (
     get_pipeline_config_for_env_name_and_config_parser
@@ -28,14 +18,6 @@ LOGGER = logging.getLogger(__name__)
 
 class S3CsvEnvironmentVariables:
     CONFIG_FILE_PATH = 'S3_CSV_CONFIG_FILE_PATH'
-    INITIAL_S3_FILE_LAST_MODIFIED_DATE = 'INITIAL_S3_FILE_LAST_MODIFIED_DATE'
-
-
-def get_default_initial_s3_last_modified_date():
-    return os.getenv(
-        S3CsvEnvironmentVariables.INITIAL_S3_FILE_LAST_MODIFIED_DATE,
-        DEFAULT_INITIAL_S3_FILE_LAST_MODIFIED_DATE
-    )
 
 
 def get_multi_csv_pipeline_config() -> MultiS3CsvConfig:
@@ -43,50 +25,6 @@ def get_multi_csv_pipeline_config() -> MultiS3CsvConfig:
         S3CsvEnvironmentVariables.CONFIG_FILE_PATH,
         MultiS3CsvConfig
     )
-
-
-def etl_new_csv_files(data_config: S3BaseCsvConfig):
-    obj_pattern_with_latest_dates = get_stored_state(
-        data_config,
-        get_default_initial_s3_last_modified_date()
-    )
-    LOGGER.info('obj_pattern_with_latest_dates: %r', obj_pattern_with_latest_dates)
-    new_s3_files = list(iter_sorted_new_s3_files_to_process(
-        obj_pattern_with_latest_dates=obj_pattern_with_latest_dates,
-        s3_bucket_name=data_config.s3_bucket_name
-    ))
-    if not new_s3_files:
-        LOGGER.info('No new file found and skipped the task.')
-        return
-    for matching_file_metadata_with_object_pattern in new_s3_files:
-        matching_file_metadata = matching_file_metadata_with_object_pattern.file_metadata
-        etag = get_s3_object_etag(
-            bucket=matching_file_metadata.bucket,
-            object_key=matching_file_metadata.name
-        )
-        LOGGER.info(
-            'ETag for s3://%s/%s is %r',
-            matching_file_metadata.bucket,
-            matching_file_metadata.name,
-            etag
-        )
-        record_import_timestamp_as_string = get_current_timestamp_as_string()
-        object_key_pattern = matching_file_metadata_with_object_pattern.object_key_pattern
-        transform_load_data(
-            matching_file_metadata.name,
-            data_config,
-            record_import_timestamp_as_string,
-        )
-        updated_obj_pattern_with_latest_dates = update_object_latest_dates(
-            obj_pattern_with_latest_dates=obj_pattern_with_latest_dates,
-            object_pattern=object_key_pattern,
-            file_modified_timestamp=matching_file_metadata.last_modified
-        )
-        upload_s3_object_json(
-            obj_pattern_with_latest_dates=updated_obj_pattern_with_latest_dates,
-            statefile_s3_bucket=data_config.state_file_bucket_name,
-            statefile_s3_object=data_config.state_file_object_name
-        )
 
 
 def csv_etl(data_pipeline_id: str, **_kwargs):
